About to make protocol.c compile.
We needed to patch ndis.h that comes with mingw (cygwin) a bit:

Do not (re-)declare enum _NDIS_REQUEST_TYPE. This is also in
ntddndis.h (also included by ndis.h) and is commented out.

Guard definition of NDIS_PROTOCOL_MAX and NDIS_PROTOCOL_MASK
Those are also defined in ntddndis.h

A comma was missing in the prototype of NdisMWanIndicateReceiveComplete().
This showed that including ndis.h probably never worked correctly.