Compile 64-bit with self signed cerificate.
Windows needs, even in test mode, a signed sys file. We
took the certificate from our WinDRBD driver (self signed).
For concenience we check the certificate into the repo.
Unfortunately this adds a build dependency to the Ewdk
(Users must call signtool which comes with the Ewdk),
one day we probably replace this by an Open Source tool
that comes with Cygwin.