Unsetting the 'reserved' (0x4000000) bit in reply tag.
Now everything works as expected (except sizes are reported
as zero)

aoe scan reports the aoe targets on the network.
aoe mount <client-MAC> creates a new Windows device
The drivers for the new disk device are installed (the
disk is handled by the disk.sys driver with partmgr.sys
filter driver stacked over it)
Last but not least the disk shows up (as unintialized) in
the partition manager which is very good. We can leave
the whole mount point management to the disk.sys/partmgr.sys
drivers in the future.